requirements.txt
awsebcli==3.7.7
blessed==1.9.5
boto==2.40.0
botocore==1.4.37
cement==2.8.2
colorama==0.3.7
coverage==4.1
Django==1.9.7
django-allauth==0.25.2
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-filter==0.13.0
django-jsonfield==1.0.0
django-jsonify==0.3.0
django-multiselectfield==0.1.3
django-notify-x==0.1.6
django-pipeline==1.6.8
django-storages==1.4.1
django-taggit==0.20.1
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
gunicorn==19.6.0
jmespath==0.9.0
jsonify==0.5
oauthlib==1.1.2
pathspec==0.3.4
Pillow==3.3.0
pinax-notifications==3.0.1
psycopg2==2.6.2
Pygments==2.1.3
python-dateutil==2.5.3
python-openid==2.2.5
PyYAML==3.11
requests==2.9.1
requests-oauthlib==0.6.1
semantic-version==2.5.0
six==1.10.0
texttable==0.8.4
wcwidth==0.1.6
websocket-client==0.37.0
